+++ src/backend/joanie/core/filters/client/certificate.py
@@ -2,8 +2,6 @@
 
 from typing import List
 
-from django.db.models import Q
-
 from django_filters import rest_framework as filters
 
 from joanie.core import enums, models
@@ -53,13 +51,12 @@ class CertificateViewSetFilter(filters.FilterSet):
         if value == enums.CERTIFICATE_ORDER_TYPE:
             # Retrieve all certificates that belong to orders of the user
             # and also legacy degrees linked to an enrollment
-            return queryset.filter(
-                Q(order__owner__username=username)
-                | (
-                    Q(enrollment__user__username=username)
-                    & Q(certificate_definition__template=enums.DEGREE)
-                )
+            enrollment_degrees = queryset.filter(
+                enrollment__user__username=username,
+                certificate_definition__template=enums.DEGREE,
             )
+            order_certificates = queryset.filter(order__owner__username=username)
+            return enrollment_degrees.union(order_certificates)
 
         if value == enums.CERTIFICATE_ENROLLMENT_TYPE:
             return queryset.filter(