+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to
   paid
 - Support of payment_schedule for certificate products
 - Display payment schedule in contract template
+- Add index on `template` field of the `CertificateDefinition` model
 
 ### Changed
 
@@ -35,6 +36,7 @@ and this project adheres to
 - Deprecated field `has_consent_to_terms` for `Order` model
 - Move signature fields before appendices in contract definition template
 - Update `handle_notification` signature backend to confirm signature 
+- Improve performance of the certificate client API endpoint
 
 ### Fixed